// src/services/workspace/types/workspace-permission.ts
export interface IWorkspacePermission {
    canCreateBoard: boolean;
    canDeleteBoard: boolean;
    canInviteMember: boolean;
    canDeleteMember: boolean;
    canChangeRole: boolean;
    canEditWorkspace: boolean;
    canDeleteWorkspace: boolean;
    canEditBoard: boolean;
    canCreateColumn: boolean;
    canDeleteColumn: boolean;
    canEditColumn: boolean;
    canCreateTask: boolean;
    canDeleteTask: boolean;
    canEditTask: boolean;
}